requirements.txt
pyspark==3.5.1
delta-spark==3.1.0
azure-eventhub==5.11.7
